Version 1.1 - Fully functional, No FreeRtos, May improve

+++ Digital_Signal_Processing.h
@@ -15,11 +15,11 @@
 #include "timers.h"
 
 #define DAC_QUEUE_LENGTH	20
-#define DAC_QUEUE_ITEM_SIZE	sizeof(double)
+#define DAC_QUEUE_ITEM_SIZE	sizeof(float)
 
 void DSP_task ();
-void DSP_Amplitude (double * x_n, double Amplitude_factor);
-void DSP_Filter (double * x_n, double *y_n, const double *h_n, double Amplitude_factor);
-void DSP_Add (double * y_n, double * y_n_1, double *y_n_2);
+void DSP_Amplitude (float * x_n, float Amplitude_factor);
+void DSP_Filter (float * x_n, float *y_n, const float *h_n, float Amplitude_factor);
+void DSP_Add (float * y_n, float * y_n_1, float *y_n_2);
 
 #endif /* SOURCE_DIGITAL_SIGNAL_PROCESSING_H_ */
